% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationCoef.R
\name{correlationCoef}
\alias{correlationCoef}
\title{function takes in 2 vectors, calculates the population correlation coefficient between the 2}
\usage{
correlationCoef(vec1, vec2)
}
\arguments{
\item{vec1}{first vector of data}

\item{vec2}{second vector of data}
}
\description{
function takes in 2 vectors, calculates the population correlation coefficient between the 2
}
\examples{
correlationCoef(vec1, vec2)
}
