% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressionCoef.R
\name{regressionCoef}
\alias{regressionCoef}
\title{finds the regression coefficient for a simple linear regression}
\usage{
regressionCoef(vec1, vec2)
}
\arguments{
\item{vec1}{first vector of data}

\item{vec2}{second vector of data}
}
\description{
finds the regression coefficient for a simple linear regression
}
\examples{
regressionCoef(vec1, vec2)
}
