% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardDev.R
\name{standardDev}
\alias{standardDev}
\title{function takes in a single vector, and returns the population standard deviation of this}
\usage{
standardDev(vec)
}
\arguments{
\item{vec}{vector of numbers}
}
\value{
the standard deviation of that set
}
\description{
function takes in a single vector, and returns the population standard deviation of this
}
\examples{
standardDev(vec)
}
